==> 02 Scripts/0. Useful functions.R <==
########################################################################
# Title: 0. Useful functions
# Author: Lea Roeller
# Date: 18/10/2025
#######################################################################



## Setting the colours for graphics ----
# main purpose here is to get a general overview of the preliminary trend
# I therefore use colourful & colour-blind friendly colours

graphic_colours <- c(
  "#E69F00", # orange
  "#56B4E9", # sky blue
  "#009E73", # bluish green
  "#F0E442", # yellow
  "#0072B2", # blue
  "#D55E00", # vermillion
  "#CC79A7", # reddish purple
  # "#999999", # grey
  "#66CCEE", # light blue
  "#882255", # dark pink/purple
  "#44AA99", # teal
  "#DDCC77", # mustard
  "#CC6677", # rose
  "#AA4499", # violet
  "#114477", # dark blue
  "#4477AA", # medium blue
  "#77AADD", # light blue
  "#117733", # dark green
  "#44BB99", # sea green
  "#88CCAA"  # pale green
)
